+++ CHANGES.md
@@ -1,6 +1,9 @@
-- Support for language
-- Support for template file in `partials` folder
-- Support for comments
+0.3
+- Added support for image alignment and captions
+- Improved plural post header image
+- Added support for language
+- Added support for template file in `partials` folder
+- Added support for comments
 
 0.2
 - Created almost whole website :)